moved messagesParser into commentsParser class
