+++ src/main/java/com/urswolfer/gerrit/client/rest/http/changes/CommentsParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2014 Urs Wolfer
+ * Copyright 2013-2020 Urs Wolfer
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 package com.urswolfer.gerrit.client.rest.http.changes;
 
 import com.google.common.reflect.TypeToken;
+import com.google.gerrit.extensions.common.ChangeMessageInfo;
 import com.google.gerrit.extensions.common.CommentInfo;
 import com.google.gerrit.extensions.common.RobotCommentInfo;
 import com.google.gson.Gson;
@@ -28,11 +29,14 @@ import java.util.List;
 import java.util.TreeMap;
 
 /**
+ * Parser for comments, messages, and robot comments.
+ *
  * @author Thomas Forrer
  */
 public class CommentsParser {
     private static final Type COMMENT_TYPE = new TypeToken<TreeMap<String, List<CommentInfo>>>() {}.getType();
     private static final Type ROBOT_COMMENT_TYPE = new TypeToken<TreeMap<String, List<RobotCommentInfo>>>() {}.getType();
+    private static final Type TYPE = new TypeToken<List<ChangeMessageInfo>>() {}.getType();
 
     private final Gson gson;
 
@@ -54,5 +58,13 @@ public class CommentsParser {
 
     public RobotCommentInfo parseSingleRobotCommentInfo(JsonObject result) {
         return gson.fromJson(result, RobotCommentInfo.class);
+    }
+
+    public List<ChangeMessageInfo> parseChangeMessageInfos(JsonElement result) {
+        return gson.fromJson(result, TYPE);
+    }
+
+    public ChangeMessageInfo parseSingleChangeMessageInfo(JsonObject result) {
+        return gson.fromJson(result, ChangeMessageInfo.class);
     }
 }
