+++ src/test/java/com/urswolfer/gerrit/client/rest/http/changes/ChangesRestClientTest.java
@@ -88,7 +88,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser, null, null,
             null, null, null, null,
-            null, null, null, null, null);
+            null, null, null, null);
 
         Changes.QueryRequest queryRequest = changes.query();
         testCase.queryParameter.apply(queryRequest).get();
@@ -106,7 +106,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null);
+            null, null, null);
         changesRestClient.query("is:open").get();
 
         EasyMock.verify(gerritRestClient);
@@ -120,7 +120,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null);
+            null, null, null);
 
         ChangeApi changeApi = changesRestClient.id(123);
 
@@ -135,7 +135,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null);
+            null, null, null);
 
         ChangeApi changeApi = changesRestClient.id("packages/test", 123);
 
@@ -150,7 +150,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null,  null, null, null, null,
-            null, null, null, null);
+            null, null, null);
 
         ChangeApi changeApi = changesRestClient.id("packages/test", "master", "Ieabd72e73f3da0df90fd6e8cba8f6c5dd7d120df");
 
@@ -165,7 +165,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser,null, null,
             null, null, null, null, null,
-            null, null, null, null);
+            null, null, null);
 
         changes.query().get();
 
@@ -186,7 +186,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser, null, null,
             null, null, null, null, null,
-            null, null, null, null);
+            null, null, null);
 
         ChangeApi changeApi = changes.create(changeInput);
